done 2.1-2.2

+++ mainFunctions2.hpp
@@ -10,4 +10,4 @@ typedef CGAL::Polygon_2<K> Polygon;
 
 
 void localSearch(Polygon* , int , int , int , int*,int );
-void simulated_annealing(Polygon* , int , int , int* ,int , int , int , int);
+void simulated_annealing(Polygon* , int , double , int* ,int , int , int , int);
